refactor: extract api into its own file
